Things fixed
Now players follow each other when they are not the active player. The speed at which they do so is controlled with the "nonActiveModifer" float found on the "PointAndClick" script on each character

The interaction now works betwen you and the golem.

The characters now face the correct way when moving, but its a bit glitchy when they reach their target position.